solution/backend/requirements.txt
uvicorn==0.35.0
fastapi==0.116.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14
pandas==2.3.1
python-dotenv==1.1.1